move safe_await_external -> external::safe_await. move set_default_executor -> external::set_default_executor.
cleanup some headers include structure
