+++ include/tmc/all_headers.hpp
@@ -11,6 +11,7 @@
 #include "tmc/aw_yield.hpp"     // IWYU pragma: export
 #include "tmc/ex_braid.hpp"     // IWYU pragma: export
 #include "tmc/ex_cpu.hpp"       // IWYU pragma: export
+#include "tmc/external.hpp"     // IWYU pragma: export
 #include "tmc/spawn_func.hpp"   // IWYU pragma: export
 #include "tmc/spawn_many.hpp"   // IWYU pragma: export
 #include "tmc/spawn_task.hpp"   // IWYU pragma: export
