+++ include/tmc/detail/ex_cpu.ipp
@@ -6,6 +6,7 @@
 // Implementation definition file for tmc::ex_cpu. This will be included
 // anywhere TMC_IMPL is defined. If you prefer to manually separate compilation
 // units, you can instead include this file directly in a CPP file.
+
 #include "tmc/detail/qu_lockfree.hpp"
 #include "tmc/detail/thread_layout.hpp"
 #include "tmc/ex_cpu.hpp"
