+++ include/tmc/spawn_func.hpp
@@ -9,6 +9,7 @@
 #include "tmc/detail/mixins.hpp"
 #include "tmc/detail/thread_locals.hpp"
 #include "tmc/task.hpp"
+
 #include <cassert>
 #include <coroutine>
 #include <functional>
