+++ include/tmc/spawn_task.hpp
@@ -10,8 +10,10 @@
 #include "tmc/detail/mixins.hpp"
 #include "tmc/detail/thread_locals.hpp"
 #include "tmc/task.hpp"
+
 #include <cassert>
 #include <coroutine>
+
 namespace tmc {
 
 template <typename Result> class aw_spawned_task_impl;
