+++ include/tmc/task.hpp
@@ -7,6 +7,7 @@
 
 #include "tmc/detail/concepts.hpp" // IWYU pragma: keep
 #include "tmc/detail/thread_locals.hpp"
+
 #include <atomic>
 #include <cassert>
 #include <coroutine>
